requirements.txt
unitypackage-extractor
pyyaml
tqdm